graph-ql/requirements.txt
ariadne==0.19.1
fastapi==0.95.1
uvicorn==0.21.1
python-dotenv==1.0.0
#selenium==4.0.0
selenium==4.19.0
beautifulsoup4==4.11.1
tldextract==3.1.2
SPARQLWrapper==2.0.0
#webdriver-manager==3.9.1
webdriver-manager==4.0.1
yfinance==0.2.37
httpx-html
lxml-html-clean==0.1.1
pdfplumber==0.11.0
Jinja2==3.0.3
xhtml2pdf==0.2.15
python-multipart